Extract MiniMax token-plan fetch into submodule
